move endpoint mocks to new directory
